@@ -8,6 +8,7 @@
     "stop": "npm run storybook-down",
     "lint": "npm run stylelint && npm run eslint",
     "test": "npm run stylelint && npm run eslint",
+    "jest": "jest",
     "build": "cross-env MODE=build NODE_ENV=production docker-compose --file storybook-compose.yml run --rm builder",
     "logs": "docker-compose --file storybook-compose.yml logs -f",
     "docker-storybook": "docker-compose --file storybook-compose.yml build",
@@ -21,6 +22,7 @@
     "stylelint-library-report": "npx stylelint --custom-formatter 'node_modules/stylelint-junit-formatter' 'pattern-library/**/*.scss' --config pattern-library/src/scss/stylelint.config.js --fix > stylelint-library-results.xml",
     "eslint-theme": "npx eslint './js/*.js' --fix -f junit -o eslint-theme-results.xml",
     "eslint-library": "npx eslint './pattern-library/src/*.js' --fix -f junit -o eslint-library-results.xml",
+    "eslint-vue": "npx eslint './vue/**/*.vue'",
     "eslint-components": "npx eslint './pattern-library/components/**/*.js' --fix -f junit -o eslint-components-results.xml",
     "eslint": "npm run eslint-theme && npm run eslint-library && npm run eslint-components",
     "builder-up": "docker-compose --file storybook-compose.yml up -d builder",
@@ -68,7 +70,7 @@
     "@ayctor/laravel-mix-svg-sprite": "^1.0.0",
     "@babel/core": "^7.13.13",
     "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/preset-env": "^7.13.12",
+    "@babel/preset-env": "^7.18.2",
     "@babel/preset-react": "^7.13.13",
     "@storybook/eslint-config-storybook": "^3.0.0",
     "@tailwindcss/forms": "^0.3.2",
@@ -78,12 +80,13 @@
     "@wordpress/eslint-plugin": "^9.0.1",
     "autoprefixer": "^10.0.4",
     "babel-eslint": "^10.1.0",
+    "babel-jest": "^28.1.1",
     "babel-loader": "^8.1.0",
     "browser-sync": "^2.27.10",
     "browser-sync-webpack-plugin": "^2.3.0",
     "cross-env": "7.0.3",
     "dotenv-webpack": "^7.1.0",
-    "eslint": "^7.23.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
@@ -92,6 +95,8 @@
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.23.1",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-vue": "^7.20.0",
+    "eslint-plugin-vuejs-accessibility": "^1.2.0",
     "faker": "^5.1.0",
     "glob": "^7.1.6",
     "jest": "^26.6.3",
