@@ -2,34 +2,37 @@
   <nav>
     <ul class="Tabs">
       <li
-        v-for="(tabName, index) in this.tabNames"
+        v-for="(tabName, index) in tabNames"
+        ref="items"
         :key="index"
         class="text-lg Tabs__tab Tab py-1 px-10"
         :class="{ 'activeTab': currentTab === tabName }"
-        ref="items"
       >
         <button @click="$emit('nav-click', tabName)">{{ tabName }}</button>
       </li>
-      <li v-if="this.dropdownTabs.length" class="px-5">
-        <button @click="toggleDropdown" class="whitespace-nowrap">
+      <li v-if="dropdownTabs.length" class="px-5">
+        <button class="whitespace-nowrap" @click="toggleDropdown">
           More
           <span
             class="material-icons-sharp text-base align-middle"
             :style="{
-              transform: this.dropdownOpen ? 'rotate(90deg)' : 'rotate(0deg)',
-              transitionProperty: 'transform',
+              transform: dropdownOpen ? 'rotate(90deg)' : 'rotate(0deg)',
+              transitionProperty: 'transform, top',
               transitionDuration: '0.2s',
+              position: 'relative',
+              top: dropdownOpen ? '0px' : '0px',
             }"
             >pending</span
           >
         </button>
         <ul
           :class="[
-            this.dropdownOpen ? 'block' : 'hidden',
+            dropdownOpen ? 'block' : 'hidden',
             'absolute',
             'border',
             'bg-white',
             'dynamic-responsive-dropdown',
+            'z-10',
           ]"
           style="padding: 10px; top: 39px; right: 0"
         >
@@ -39,8 +42,9 @@
             :class="{ 'active': currentTab === dropdownTab }"
           >
             <a
-              @click="$emit('nav-click', dropdownTab)"
               class="cursor-pointer inline-block w-full"
+              @keyup.enter="$emit('nav-click', dropdownTab)"
+              @click="$emit('nav-click', dropdownTab)"
               >{{ dropdownTab }}</a
             >
           </li>
@@ -50,13 +54,15 @@
   </nav>
 </template>
 <script>
-import _debounce from 'lodash/debounce';
-import _remove from 'lodash/remove';
+import { fillTabs } from '../helpers/_helpers.js';
 
 export default {
-  props: ['currentTab'],
+  props: {
+    currentTab: { type: String, required: true },
+  },
   data() {
     return {
+      windowPreviousWidth: 0,
       tabNames: [
         'Stories',
         'Media Coverage',
@@ -69,12 +75,13 @@ export default {
     };
   },
   created() {
-    window.addEventListener('resize', _debounce(this.responsiveTabs, 100));
+    window.addEventListener('resize', this.responsiveTabs);
   },
-  destroyed() {
-    window.removeEventListener('resize', _debounce(this.responsiveTabs, 100));
+  unmounted() {
+    window.removeEventListener('resize', this.responsiveTabs);
   },
   mounted() {
+    this.windowPreviousWidth = window.innerWidth;
     this.calculateTabs();
     this.responsiveTabs();
   },
@@ -89,35 +96,21 @@ export default {
         });
       });
     },
-    responsiveTabs(e) {
-      let newTabs = [];
+    responsiveTabs() {
+      // let newTabs = [];
 
       // if window smaller than width of ul on desktop
       if (window.innerWidth < 1066) {
-        this.tabs.forEach((item, i) => {
-          // current tab name
-          let label = Object.keys(item)[0];
+        let tabs = fillTabs(
+          window.innerWidth,
+          this.tabs,
+          this.tabNames,
+          this.dropdownTabs,
+          this.windowPreviousWidth
+        );
 
-          // width of responsive dynamic dropdown
-          let amt = 106;
-
-          if (i === 0) {
-            amt = 0;
-          }
-
-          // if window is smaller than the right edge of the tab
-          if (window.innerWidth - amt < item[label].right) {
-            if (!this.dropdownTabs.includes(label)) {
-              this.dropdownTabs.push(label);
-              _remove(this.tabNames, (tabname) => tabname === label);
-            }
-          } else {
-            newTabs.push(label);
-            _remove(this.dropdownTabs, (tabname) => tabname === label);
-          }
-
-          this.tabNames = newTabs;
-        });
+        this.tabNames = tabs.tabNames;
+        this.dropdownTabs = tabs.dropdownTabs;
       } else {
         // reset default tab state
         this.tabNames = [
@@ -128,6 +121,8 @@ export default {
         ];
         this.dropdownTabs = [];
       }
+
+      this.windowPreviousWidth = window.innerWidth;
     },
   },
 };
