@@ -1,9 +1,7 @@
 <template>
   <div>
-    <h2 class="pb-8 text-lg">
-      <b>No results found for "{{ query }}".</b>
-    </h2>
-    <h2 class="pb-8">Recommended for you</h2>
+    <p class="pb-8">No results found for "{{ query }}".</p>
+    <h2 class="pb-8 text-lg"><b>Recommended for you</b></h2>
     <ais-index
       index-name="prod_news_searchable_posts"
       index-id="noresult-videos"
@@ -16,7 +14,7 @@
       <div class="wp-block">
         <div class="wp-block nc-slider-with-teaser-pair">
           <ais-hits>
-            <template v-slot="{ items, sendEvent }">
+            <template v-slot="{ items }">
               <div class="mb-10">
                 <div class="grid md:grid-cols-2 gap-8">
                   <div v-for="item in items.slice(0, 2)" :key="item.objectID">
@@ -27,6 +25,7 @@
                             <img
                               :src="item.images.teaser_new.url"
                               class="hover:brightness-90 transition ease-in-out duration-300"
+                              alt="`${item.post_title}: ${item.summary}`"
                             />
                           </a>
                         </div>
@@ -78,6 +77,7 @@
                                   :src="item.images.teaser_new.url"
                                   class="attachment-teaser_new size-teaser_new hover:brightness-90 transition ease-in-out duration-300"
                                   sizes="(max-width: 1080px) 100vw, 1080px"
+                                  :alt="`${item.post_title}: ${item.summary}`"
                                 />
                               </a>
                             </div>
@@ -146,7 +146,11 @@
 <script>
 export default {
   props: {
-    query: String,
+    query: {
+      type: String,
+      default: '',
+      required: true,
+    },
   },
   methods: {
     getTeaserPair(items) {
