@@ -2,7 +2,7 @@
   <div>
     <ais-pagination @page-change="pageChange">
       <template
-        v-slot="{
+        #default="{
           currentRefinement,
           nbPages,
           isFirstPage,
@@ -20,17 +20,39 @@
           <li class="flex justify-center items-center">
             <a
               :href="createURL(currentRefinement - 1)"
+              :disabled="isFirstPage ? true : false"
+              :class="[isFirstPage ? 'pointer-events-none' : '', 'mr-2']"
+              :style="{ color: isFirstPage ? '#8b8b8b' : 'black' }"
               @click.prevent="refine(currentRefinement - 1)"
-              :style="{ color: isFirstPage ? '#D8D8D8' : 'black' }"
             >
-              <span class="material-icons-sharp chevron"> chevron_left </span>
+              <span
+                :class="[
+                  'material-icons-sharp',
+                  isFirstPage
+                    ? 'text-black text-xl leading-none border border-black rounded'
+                    : 'text-white text-xl leading-none border border-black rounded bg-black hover:text-black hover:bg-white',
+                ]"
+              >
+                chevron_left
+              </span>
             </a>
             <a
               :href="createURL(currentRefinement + 1)"
+              :disabled="isLastPage ? true : false"
+              :class="[isLastPage ? 'pointer-events-none' : '']"
+              :style="{ color: isLastPage ? '#8b8b8b' : 'black' }"
               @click.prevent="refine(currentRefinement + 1)"
-              :style="{ color: isLastPage ? '#D8D8D8' : 'black' }"
             >
-              <span class="material-icons-sharp chevron"> chevron_right </span>
+              <span
+                :class="[
+                  'material-icons-sharp',
+                  isLastPage
+                    ? 'text-black text-xl leading-none border border-black rounded'
+                    : 'text-white text-xl leading-none border border-black rounded bg-black hover:text-black hover:bg-white',
+                ]"
+              >
+                chevron_right
+              </span>
             </a>
           </li>
         </ul>
@@ -48,7 +70,7 @@ export default {
     };
   },
   methods: {
-    pageChange(...args) {
+    pageChange() {
       const element = document.getElementById('site-search-searchbox');
       element.scrollIntoView({ behavior: 'smooth', block: 'end' });
     },
