@@ -1,3 +1,52 @@
+<script>
+/* eslint-disable vuejs-accessibility/anchor-has-content */
+/* eslint-disable vuejs-accessibility/heading-has-content */
+import Pagination from '../Pagination.vue';
+import NoResults from '../NoResults.vue';
+import { createWidgetMixin } from 'vue-instantsearch/vue3/es';
+
+const connector = (renderFn, unmountFn) => () => ({
+  init() {
+    renderFn({ hasResult: true }, true);
+  },
+
+  render({ scopedResults, helper }) {
+    const hasResult =
+      scopedResults &&
+      scopedResults.find(
+        (indexResult) =>
+          indexResult.indexId === 'videos' && indexResult.results.nbHits > 0
+      );
+
+    renderFn(
+      {
+        hasResult,
+        query: helper.state.query,
+      },
+      false
+    );
+  },
+
+  dispose() {
+    unmountFn();
+  },
+});
+export default {
+  components: {
+    Pagination,
+    NoResults,
+  },
+  mixins: [createWidgetMixin({ connector })],
+  props: {
+    currentTab: { type: String, required: true },
+  },
+  data() {
+    return {
+      hover: null,
+    };
+  },
+};
+</script>
 <template>
   <div
     v-show="currentTab === 'Videos'"
@@ -9,18 +58,26 @@
         <ais-index index-name="prod_news_videos" index-id="videos">
           <ais-configure :hits-per-page.camel="5" />
           <ais-hits>
-            <template v-slot="{ items, sendEvent }">
+            <template #default="{ items, sendEvent }">
               <ul>
                 <li v-for="item in items" :key="item.objectID">
                   <div
-                    class="!flex !flex-row pb-8 mb-12 border-b border-gray-700"
+                    class="flex flex-col md:flex-row pb-8 mb-12 border-b border-gray-700"
                   >
                     <div
-                      class="!w-1/4 !m-0 !p-0 transition ease-in-out duration-300"
-                      :class="{
-                        'brightness-90': this.hover === item.objectID,
-                      }"
                       v-if="item.thumbnail"
+                      class=""
+                      :class="[
+                        hover === item.objectID ? 'brightness-90' : '',
+                        'w-full',
+                        'mt-4',
+                        'md:w-1/4',
+                        '!m-0',
+                        '!p-0',
+                        'transition',
+                        'ease-in-out',
+                        'duration-300',
+                      ]"
                     >
                       <a
                         :href="`https://www.youtube.com/watch?v=${item.objectID}`"
@@ -28,28 +85,32 @@
                         @click="sendEvent('click', item, 'Video Clicked')"
                       >
                         <img
+                          :alt="item.title"
                           class="!object-cover"
                           :src="item.thumbnail.url"
                           @mouseover="hover = item.objectID"
+                          @focus="hover = item.objectID"
                           @mouseleave="hover = null"
-                          :alt="item.title"
+                          @blur="hover = null"
                         />
                       </a>
                     </div>
-                    <div class="w-3/4">
+                    <div class="w-full mt-4 md:mt-0 md:w-3/4 md:ml-4">
                       <a
                         :href="`https://www.youtube.com/watch?v=${item.objectID}`"
                         :class="{
-                          'text-link-hover': this.hover === item.objectID,
+                          'text-link-hover': hover === item.objectID,
                         }"
-                        @mouseover="hover = item.objectID"
-                        @mouseleave="hover = null"
-                        @click="sendEvent('click', item, 'Video Clicked')"
                         :target="'_blank'"
+                        @mouseover="hover = item.objectID"
+                        @focus="hover = item.objectID"
+                        @mouseleave="hover = null"
+                        @blur="hover = null"
+                        @click="sendEvent('click', item, 'Video Clicked')"
                       >
                         <h2
                           :class="[
-                            'thumbnail' in item ? 'pl-6' : 'pl-0',
+                            'thumbnail' in item ? 'pl-0 md:pl-6' : 'pl-0',
                             'font-bold',
                             'text-base',
                             'mb-1.5',
@@ -60,7 +121,7 @@
 
                         <p
                           :class="[
-                            'thumbnail' in item ? 'pl-6' : 'pl-0',
+                            'thumbnail' in item ? 'pl-0 md:pl-6' : 'pl-0',
                             'font-sans',
                             'text-base',
                           ]"
@@ -73,7 +134,9 @@
               </ul>
             </template>
           </ais-hits>
-          <pagination></pagination>
+          <div v-show="state.hasResult">
+            <pagination></pagination>
+          </div>
         </ais-index>
       </div>
       <div v-show="!state.hasResult && state.query">
@@ -82,50 +145,3 @@
     </div>
   </div>
 </template>
-<script>
-import Pagination from '../components/Pagination.vue';
-import NoResults from '../components/NoResults.vue';
-import { createWidgetMixin } from 'vue-instantsearch/vue3/es';
-
-const connector =
-  (renderFn, unmountFn) =>
-  (widgetParams = {}) => ({
-    init({}) {
-      renderFn({ hasResult: true }, true);
-    },
-
-    render({ scopedResults, helper }) {
-      const hasResult =
-        scopedResults &&
-        scopedResults.find(
-          (indexResult) =>
-            indexResult.indexId === 'videos' && indexResult.results.nbHits > 0
-        );
-
-      renderFn(
-        {
-          hasResult,
-          query: helper.state.query,
-        },
-        false
-      );
-    },
-
-    dispose() {
-      unmountFn();
-    },
-  });
-export default {
-  components: {
-    Pagination,
-    NoResults,
-  },
-  props: ['currentTab'],
-  data() {
-    return {
-      hover: null,
-    };
-  },
-  mixins: [createWidgetMixin({ connector })],
-};
-</script>
